refactor(OMN-1867): Rename mixin file to match class name per CLAUDE.md
Rename mixin_postgres_error_handler.py to mixin_postgres_error_response.py
to comply with CLAUDE.md naming convention requiring exact correspondence
between file name and class name (mixin_<name>.py → Mixin<Name>).
